Added users should have valid email addresses, random password, and disabled
It is important to not create non-secure user accounts as part of the csv
import.  Before this change an account was created with empty email,
and user name as the password.  The change sets the email to a valid address,
sets a random password, and creates the account as disabled.  Administrators
can then go an re-enable / reset password if needed.